requirements.txt
mysql.connector
mysql-connector-python
colorama
multiset

# Need to have mysql and phpmyadmin installed on system